requirements.txt
async-timeout==4.0.3
python-dotenv>=1.0.0
numpy==1.26.4
pandas==2.2.1
pymodbus==3.6.4
pyModbusTCP==0.2.1
python-dateutil==2.9.0
pytz==2024.1
redis==5.0.3
six==1.16.0
tzdata==2024.1